Refactored thank you screen
